+++ android/src/main/java/ng/com/piccmaq/flutter/flutter_mobile_vision_2/util/MobileVisionException.java
@@ -1,0 +1,12 @@
+package ng.com.piccmaq.flutter.flutter_mobile_vision_2.util;
+
+public class MobileVisionException extends Exception {
+
+    public MobileVisionException(String message) {
+        super(message);
+    }
+
+    public MobileVisionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}