+++ example/android/app/src/main/java/io/github/edufolly/flutter_mobile_vision_example/MainActivity.java
@@ -1,6 +1,0 @@
-package io.github.edufolly.flutter_mobile_vision_example;
-
-import io.flutter.embedding.android.FlutterActivity;
-
-public class MainActivity extends FlutterActivity {
-}