+++ example/android/app/src/main/java/ng/com/piccmaq/flutter/flutter_mobile_vision_2_example/MainActivity.java
@@ -1,0 +1,6 @@
+package ng.com.piccmaq.flutter.flutter_mobile_vision_2_example;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}