+++ example/lib/face_detail.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_mobile_vision/flutter_mobile_vision.dart';
+import 'package:flutter_mobile_vision_2/flutter_mobile_vision_2.dart';
 
 ///
 ///