+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.6] - 2021-07-13
+
+* perform non-null/null-safety test.
+
+## [0.1.5] - 2021-07-13
+
+* non-null/null-safety plugin package.
+
 ## [0.1.4+1] - 2020-04-28
 
 * Rename the plugin package.